Extract "version-checker" file
